pyproject.toml
[project]
name = "luna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "dataclasses-json>=0.6.7",
    "kokoro>=0.9.4",
    "kokoro-onnx>=0.5.0",
    "mlflow>=3.10.1",
    "openai-whisper>=20250625",
    "prompt-toolkit>=3.0.52",
    "psycopg>=3.3.3",
    "pydantic>=2.12.5",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "torch>=2.10.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
